taxmst and hsnmst added

+++ app/src/main/java/in/net/maitri/xb/billing/BillItems.java
@@ -1,6 +1,5 @@
 package in.net.maitri.xb.billing;
 
-import java.math.BigDecimal;
 
 /**
  * Created by SYSRAJ4 on 16/11/2017.
@@ -14,11 +13,11 @@ public class BillItems {
     private int item_id;
     private double net_rate;
     private double rate;
-    private BigDecimal taxRate1;
-    private BigDecimal taxRate2;
-    private BigDecimal taxAmt1;
-    private BigDecimal taxAmt2;
-    private BigDecimal taxSaleAmt;
+    private float taxRate1;
+    private float taxRate2;
+    private float taxAmt1;
+    private float taxAmt2;
+    private float taxSaleAmt;
 
 
     public BillItems () {}
@@ -42,8 +41,8 @@ public class BillItems {
     }
 
     public BillItems( int cat_id, int item_id,String desc, double qty,double net_rate,
-                      double rate, double amount, BigDecimal taxRate1, BigDecimal taxRate2,
-                      BigDecimal taxAmt1, BigDecimal taxAmt2, BigDecimal taxSaleAmt) {
+                      double rate, double amount, float taxRate1, float taxRate2,
+                      float taxAmt1, float taxAmt2, float taxSaleAmt) {
         this(cat_id, item_id, desc, qty,net_rate, rate, amount);
         this.taxAmt1 = taxAmt1;
         this.taxRate1 = taxRate1;
@@ -53,43 +52,43 @@ public class BillItems {
     }
 
 
-    public BigDecimal getTaxRate1() {
+    public float getTaxRate1() {
         return taxRate1;
     }
 
-    public void setTaxRate1(BigDecimal taxRate1) {
+    public void setTaxRate1(float taxRate1) {
         this.taxRate1 = taxRate1;
     }
 
-    public BigDecimal getTaxRate2() {
+    public float getTaxRate2() {
         return taxRate2;
     }
 
-    public void setTaxRate2(BigDecimal taxRate2) {
+    public void setTaxRate2(float taxRate2) {
         this.taxRate2 = taxRate2;
     }
 
-    public BigDecimal getTaxAmt1() {
+    public float getTaxAmt1() {
         return taxAmt1;
     }
 
-    public void setTaxAmt1(BigDecimal taxAmt1) {
+    public void setTaxAmt1(float taxAmt1) {
         this.taxAmt1 = taxAmt1;
     }
 
-    public BigDecimal getTaxAmt2() {
+    public float getTaxAmt2() {
         return taxAmt2;
     }
 
-    public void setTaxAmt2(BigDecimal taxAmt2) {
+    public void setTaxAmt2(float taxAmt2) {
         this.taxAmt2 = taxAmt2;
     }
 
-    public BigDecimal getTaxSaleAmt() {
+    public float getTaxSaleAmt() {
         return taxSaleAmt;
     }
 
-    public void setTaxSaleAmt(BigDecimal taxSaleAmt) {
+    public void setTaxSaleAmt(float taxSaleAmt) {
         this.taxSaleAmt = taxSaleAmt;
     }
